refactor: rename card list to card list-page
